mm: Fix truncated major/minor output in PID maps
The major and minor can be up to 12 bits and 20 bits long, respectively.
The current output generated after the micro-optimizations applied to
show_vma_header_prefix() only show up to 8 bits of each (2 hex digits),
resulting in the major and minor being truncated.

Fix it by introducing new optimized macros to print up to 3 and 5 hex
digits for the major and minor, respectively.

Change-Id: I3af70dd67cfa6b5fc4cb06c022b8839bd69c8a94
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>